Allow auditing of entities with many-many relations
The changes made in c619a4efaee2d1b73d1f634a61d285257a2d47cf have made
it impossible to load the history of entities that have a many to many
relationship with another entity. I think the old approach of just
providing an empty collection and the notice in the docs is preferable
to not being able to audit the entity at all. Hence I reverted the
change that throws an exception if a many to many relationship is found