fix: use the type of association fields in LogRevisionsListener
When building the "UPDATE" query in postFlush, LogRevisionsListener
computes the type of each field to update. When the field was an association
field, the type was not used in the call of Connection::executeQuery, causing
the UPDATE query to fail.

This commit simply adds the computed type in the "types" array