Fallback to native entity if no revision found for properties
If an audited entity has relational properties, the reader will
try to find them as the same version, or previous.

But if the targeted class does not has any revision at all, this
produces a NoRevisionFound exception.

This should be catched and should fallback to a classic doctrine find query.